Add import_zipcodes command to… import french zipcodes db
This database is several megabytes, and is available at a stable
URL (normaly…) from opendata source.

So we here prefer to offer tools to download an import
rather than embeding it in libreosteo repository, it is a bit less reliable,
but :

- makes the repository lighter
- makes the update easier

Maybe adapting windows package to handle it will be necessary.